requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
SQLAlchemy==2.0.38
WeasyPrint==60.1
gunicorn==21.2.0
redis==5.0.1
celery==5.3.6
psycopg2-binary==2.9.9
PyJWT==2.8.0
email-validator==2.1.0.post1
python-dateutil==2.8.2